code/requirements.txt
albumentations
matplotlib
numpy
opencv-python
opencv-python-headless
Pillow
segmentation-models-pytorch
torch
torchvision
tqdm
